validation and location_branch UI

+++ src/delete.php
@@ -12,45 +12,98 @@
     $status = false;
 
     if($_POST['type'] == "BRA"){
-        $query->setQuery("DELETE FROM LOCATION_BRANCH   
+        $query->setQuery("SELECT COMPLAINT_ID FROM COMPLAINT  
             WHERE BRANCH_ID = :id");
-
         $param = array(
             ":id" => trim($_POST['id'])
         );
+        $r = $query->fetch_array_with_param($param);
 
-        $status = $query->insertInto($param);
+        //if not use by other complaint row continue
+        if($r[0][0] == 0){
+            $query->setQuery("DELETE FROM LOCATION_BRANCH   
+            WHERE BRANCH_ID = :id");
+
+            $param = array(
+                ":id" => trim($_POST['id'])
+            );
+
+            $status = $query->insertInto($param);
         
-        $query->setQuery("DELETE FROM BRANCH   
+            $query->setQuery("DELETE FROM BRANCH   
             WHERE BRANCH_ID = :id");
 
-        $param = array(
-            ":id" => trim($_POST['id'])
-        );
+            $param = array(
+                ":id" => trim($_POST['id'])
+            );
+            $status = $query->insertInto($param);
+        } else {
+            die('DependecyErr');
+        }    
     }
 
     else if($_POST['type'] == "CAT"){
-        $query->setQuery("DELETE FROM CATEGORY  
+        $query->setQuery("SELECT COMPLAINT_ID FROM COMPLAINT  
             WHERE CATEGORY_ID = :id");
+        
         $param = array(
             ":id" => trim($_POST['id'])
         );
+        
+        $r = $query->fetch_array_with_param($param);
+
+        //if not use by other complaint row continue
+        if($r[0][0] == 0){
+            $query->setQuery("DELETE FROM CATEGORY  
+                WHERE CATEGORY_ID = :id");
+            
+            $param = array(
+                ":id" => trim($_POST['id'])
+            );
+            $status = $query->insertInto($param);
+        } else {
+            die('DependecyErr');
+        }
     }
 
     else if($_POST['type'] == "LOC"){
-        $query->setQuery("DELETE FROM LOCATION_BRANCH  
+        $query->setQuery("SELECT COMPLAINT_ID FROM COMPLAINT  
             WHERE LOCATION_ID = :id");
-        $param = array(
-            ":id" => trim($_POST['id'])
-        );
-
-        $status = $query->insertInto($param);
         
-        $query->setQuery("DELETE FROM LOCATION 
-            WHERE LOCATION_ID = :id");
         $param = array(
             ":id" => trim($_POST['id'])
         );
+        
+        $r = $query->fetch_array_with_param($param);
+
+        //if not use by other complaint row continue
+        if( empty($r) ||$r[0][0] == 0){
+            $query->setQuery("DELETE FROM LOCATION_BRANCH  
+                WHERE LOCATION_ID = :id");
+            $param = array(
+                ":id" => trim($_POST['id'])
+            );
+
+
+            $status = $query->insertInto($param);
+            $query->setQuery("DELETE FROM LOCATION 
+                WHERE LOCATION_ID = :id");
+            $param = array(
+                ":id" => trim($_POST['id'])
+            );
+            $status = $query->insertInto($param);
+        }
+        else {
+            die('DependecyErr');
+        }
+    } else if($_POST['type'] == "REL"){
+        $query->setQuery("DELETE FROM LOCATION_BRANCH  
+                WHERE LOCATION_ID = :loc_id AND BRANCH_ID = :bra_id");
+        $param = array(
+            ":loc_id" => trim($_POST['loc_id']),
+            ":bra_id" => trim($_POST['bra_id']),
+        );
+        $status = $query->insertInto($param);
     }
 
     else if($_POST['type'] == "ADM"){
@@ -59,19 +112,27 @@
 
         if(strcmp($pass,$c_pass) == 0 ){
 
-            $query->setQuery("DELETE FROM ADMIN 
-                WHERE USERNAME = :id 
-                AND ADMIN_TYPE <> 'SYS'");
-            $param = array(
-                ":id" => trim($_POST['id'])
-            );
+            $query->setQuery("SELECT ADMIN_TYPE FROM ADMIN
+                WHERE USERNAME = '".$_SESSION['username']."'");
+            
+            $access = $query->fetch_array();
+            if($access[0][0] == 'SYS'){
+
+                $query->setQuery("DELETE FROM ADMIN 
+                    WHERE USERNAME = :id 
+                    AND ADMIN_TYPE <> 'SYS'");
+                $param = array(
+                    ":id" => trim($_POST['id'])
+                );
+                $status = $query->insertInto($param);
+            } else {
+                die("accessErr");
+            }
         }
     }
 
-    if(!empty($param))
-        $status = $query->insertInto($param);
     if($status)
-        header("Location: ./admin_setting.php?status=success");
+        echo "OK";
     else
         echo "Failed to delete data from DB";
 ?>
